pc-port: native build toolchain — port compiles, links, and renders real content
The port now builds to a native x86 exe that opens a GL window and renders real
topmenu.fsys scene content through the game's own decompiled GSgfx draw path.

Build tooling (tools/):
- pcport_gen.py    flips `#if 1 asm{#include .inc} #else <C>` wrappers to their C
                   body for the PC build (matching/CodeWarrior build untouched);
                   per-file preambles for a few decls. Solved the asm-wrapper gate
                   (74 wrappers, ~25/29 HSD TUs compile).
- pcport_compat.h  PPC intrinsic shims (__fabs/__cntlzw) for the PC build only.
- pcport_link.py   compiles 33 real TUs (main + HSD resolver + 24 HSD + game
                   gfx/render) with clang -m32 -fms-compatibility, links vs the
                   CMake-built shim/glfw/glad libs, and auto-stubs the 176-symbol
                   render closure (/FORCE:MULTIPLE + /ERRORLIMIT:0); converges in 2.
- pcport_build.ps1 one-command CMake (VS18 / Win32) configure+build helper.
- pcport_probe/    standalone HSD-archive resolver probe (C + py), verified on real
                   topmenu.fsys (FSYS->LZSS->HSD parse->relocate->scene_data root).

Shim fixes (src/pcport/):
- gx_shim.h + audio/dvd/os/pad_shim.h: u32/s32 typedefs int->long to match
  include/dolphin/types.h (strict compilers reject the redefinition; PCPORT-only).
- pcport_main.c: DumpFramebufferBMP (env PCPORT_DUMP -> 24-bit BMP) for screenshots.

All PCPORT-guarded; the GameCube matching build and gates are unaffected (the
compile gates already skip src/pcport).

+++ src/pcport/dvd_shim.h
@@ -30,8 +30,8 @@ extern "C" {
 /* Standard types */
 typedef unsigned char  u8;
 typedef unsigned short u16;
-typedef unsigned int   u32;
-typedef signed int     s32;
+typedef unsigned long    u32;
+typedef signed long      s32;
 typedef signed long long s64;
 typedef int            BOOL;
